Changes:
Ruleset now have a rule code scheme.  The reason for this is twofold.
First, by having short and concise error codes it makes it easier to
specify on the command line which rules should be omitted.  Second, this
provides an identifier which can be used to lookup additional
information about a result (e.g. info, hints) that can be included in
results.

The OAS2-Enhanced and OAS2-Fortellis rulsets are nearly complete.  All
unit tests have been completed for each rule and function.

Remaining Work:
* Fix the bug in the rule wpar_f003 that should detect when references
aren't used for parameter and reponse objects.  This is complicated by
the fact the JSON schema parser makes `$ref` declarations transparent.
There are ways to work around this...

* Implement flags that specify which rulesets to use and which
individual rules have been omitted.